// src/types.ts
export type ResumeUpdater =
  | ResumeUpdater.UpdateContactIcons
  | ResumeUpdater.UpdateName
  | ResumeUpdater.UpdateSections
  | ResumeUpdater.UpdateSettings
  | ResumeUpdater.UpdateSummary;

export namespace ResumeUpdater {
  export enum Type {
    UpdateContactIcons = "me.kthomas.resume.ResumeUpdater.UpdateContactIcons",
    UpdateName = "me.kthomas.resume.ResumeUpdater.UpdateName",
    UpdateSections = "me.kthomas.resume.ResumeUpdater.UpdateSections",
    UpdateSettings = "me.kthomas.resume.ResumeUpdater.UpdateSettings",
    UpdateSummary = "me.kthomas.resume.ResumeUpdater.UpdateSummary",
  }
  
  export interface UpdateContactIcons {
    type: ResumeUpdater.Type.UpdateContactIcons;
    updater: ModifyContactItems;
  }
  
  export interface UpdateName {
    type: ResumeUpdater.Type.UpdateName;
    newName: string;
  }
  
  export interface UpdateSections {
    type: ResumeUpdater.Type.UpdateSections;
    updater: ModifySections;
  }
  
  export interface UpdateSettings {
    type: ResumeUpdater.Type.UpdateSettings;
    updater: SettingsUpdater;
  }
  
  export interface UpdateSummary {
    type: ResumeUpdater.Type.UpdateSummary;
    newSummary: string;
  }
}

export type ModifyContactItems =
  | ModifyContactItems.AddContactItem
  | ModifyContactItems.MoveContactItem
  | ModifyContactItems.RemoveContactItem
  | ModifyContactItems.UpdateContactItem;

export namespace ModifyContactItems {
  export enum Type {
    AddContactItem = "me.kthomas.resume.ModifyContactItems.AddContactItem",
    MoveContactItem = "me.kthomas.resume.ModifyContactItems.MoveContactItem",
    RemoveContactItem = "me.kthomas.resume.ModifyContactItems.RemoveContactItem",
    UpdateContactItem = "me.kthomas.resume.ModifyContactItems.UpdateContactItem",
  }
  
  export interface AddContactItem {
    type: ModifyContactItems.Type.AddContactItem;
    item: ResumeContactItem;
  }
  
  export interface MoveContactItem {
    type: ModifyContactItems.Type.MoveContactItem;
    oldIndex: number;
    newIndex: number;
  }
  
  export interface RemoveContactItem {
    type: ModifyContactItems.Type.RemoveContactItem;
    index: number;
  }
  
  export interface UpdateContactItem {
    type: ModifyContactItems.Type.UpdateContactItem;
    index: number;
    item: ResumeContactItem;
  }
}

export type ModifySections =
  | ModifySections.AddSection
  | ModifySections.MoveSection
  | ModifySections.RemoveSection
  | ModifySections.UpdateSection;

export namespace ModifySections {
  export enum Type {
    AddSection = "me.kthomas.resume.ModifySections.AddSection",
    MoveSection = "me.kthomas.resume.ModifySections.MoveSection",
    RemoveSection = "me.kthomas.resume.ModifySections.RemoveSection",
    UpdateSection = "me.kthomas.resume.ModifySections.UpdateSection",
  }
  
  export interface AddSection {
    type: ModifySections.Type.AddSection;
    section: ResumeSection;
  }
  
  export interface MoveSection {
    type: ModifySections.Type.MoveSection;
    oldIndex: number;
    newIndex: number;
  }
  
  export interface RemoveSection {
    type: ModifySections.Type.RemoveSection;
    index: number;
  }
  
  export interface UpdateSection {
    type: ModifySections.Type.UpdateSection;
    index: number;
    updater: SectionUpdater;
  }
}

export type SettingsUpdater =
  | SettingsUpdater.UpdateBaseFontSize;

export namespace SettingsUpdater {
  export enum Type {
    UpdateBaseFontSize = "me.kthomas.resume.SettingsUpdater.UpdateBaseFontSize",
  }
  
  export interface UpdateBaseFontSize {
    type: SettingsUpdater.Type.UpdateBaseFontSize;
    size: number;
  }
}

export interface ResumeContactItem {
  icon: ResumeContactIcon;
  value: string;
}

export interface ResumeSection {
  title: string;
  contentItems: ResumeSectionItem[];
}

export type SectionUpdater =
  | SectionUpdater.AddContent
  | SectionUpdater.MoveContent
  | SectionUpdater.RemoveContent
  | SectionUpdater.UpdateContent
  | SectionUpdater.UpdateTitle;

export namespace SectionUpdater {
  export enum Type {
    AddContent = "me.kthomas.resume.SectionUpdater.AddContent",
    MoveContent = "me.kthomas.resume.SectionUpdater.MoveContent",
    RemoveContent = "me.kthomas.resume.SectionUpdater.RemoveContent",
    UpdateContent = "me.kthomas.resume.SectionUpdater.UpdateContent",
    UpdateTitle = "me.kthomas.resume.SectionUpdater.UpdateTitle",
  }
  
  export interface AddContent {
    type: SectionUpdater.Type.AddContent;
    content: ResumeSectionItem;
  }
  
  export interface MoveContent {
    type: SectionUpdater.Type.MoveContent;
    oldIndex: number;
    newIndex: number;
  }
  
  export interface RemoveContent {
    type: SectionUpdater.Type.RemoveContent;
    index: number;
  }
  
  export interface UpdateContent {
    type: SectionUpdater.Type.UpdateContent;
    index: number;
    updater: SectionItemUpdater;
  }
  
  export interface UpdateTitle {
    type: SectionUpdater.Type.UpdateTitle;
    newTitle: string;
  }
}

export enum ResumeContactIcon {
  Email = "Email",
  Phone = "Phone",
  Website = "Website",
  Github = "Github",
  LinkedIn = "LinkedIn",
}